Test tag like input selections

+++ Business/Entity/Genre.cs
@@ -2,7 +2,7 @@
 
 namespace Business.Entity;
 
-public record Genre(long? Id, string Name)
+public record Genre(byte? Id, string Name)
 {
     public Genre(GenreDTO genreDto) : this(genreDto.Id, genreDto.Name) { }
     public GenreDTO ToDto() => new(Id, Name);
