+++ Business/Entity/Setting.cs
@@ -2,7 +2,7 @@
 
 namespace Business.Entity;
 
-public record Setting(long? Id, string Description)
+public record Setting(byte? Id, string Description)
 {
     public Setting(SettingDTO settingDto) : this(settingDto.Id, settingDto.Description) { }
     public SettingDTO GetDto() => new SettingDTO(Id, Description);
