+++ Business/Entity/Theme.cs
@@ -2,7 +2,7 @@
 
 namespace Business.Entity;
 
-public record Theme(long? Id, string Description)
+public record Theme(byte? Id, string Description)
 {
     public Theme(ThemeDTO themeDto) : this(themeDto.Id, themeDto.Description) { }
     public ThemeDTO GetDto() => new ThemeDTO(Id, Description);
