+++ IBDbWebApplication/Models/Entity/SettingModel.cs
@@ -2,9 +2,9 @@
 
 public class SettingModel
 {
-    public long? Id { get; set; }
+    public byte? Id { get; set; }
     public string Description { get; set; }
 
-    public SettingModel(long? id, string description) => 
+    public SettingModel(byte? id, string description) => 
         (Id, Description) = (id, description);
 }
