+++ IBDbWebApplication/Models/Entity/ThemeModel.cs
@@ -2,9 +2,9 @@
 
 public class ThemeModel
 {
-    public long? Id { get; set; }
+    public byte? Id { get; set; }
     public string Description { get; set; }
     
-    public ThemeModel(long? id, string description) => 
+    public ThemeModel(byte? id, string description) => 
         (Id, Description) = (id, description);
 }
