+++ Interface/DTO/SettingDTO.cs
@@ -1,3 +1,3 @@
 ﻿namespace Interface.DTO;
 
-public record struct SettingDTO(long? Id, string Description);
+public record struct SettingDTO(byte? Id, string Description);
